move ctcs sorting and violation code into `augmentation` dir
